+++ src/siun/providers/__init__.py
@@ -1,3 +1,3 @@
-from .base import *
-from .generic import *
-from .pacman import *
+from .base import UpdateProvider as UpdateProvider
+from .generic import UpdateProviderGeneric as UpdateProviderGeneric
+from .pacman import UpdateProviderPacman as UpdateProviderPacman